split SimpleTeleport into it's own file
