+++ OpenRA.Mods.RA/Activities/SimpleTeleport.cs
@@ -1,0 +1,28 @@
+#region Copyright & License Information
+/*
+ * Copyright 2007-2014 The OpenRA Developers (see AUTHORS)
+ * This file is part of OpenRA, which is free software. It is made
+ * available to you under the terms of the GNU General Public License
+ * as published by the Free Software Foundation. For more information,
+ * see COPYING.
+ */
+#endregion
+
+using OpenRA.Traits;
+
+namespace OpenRA.Mods.RA.Activities
+{
+	public class SimpleTeleport : Activity
+	{
+		CPos destination;
+
+		public SimpleTeleport(CPos destination) { this.destination = destination; }
+
+		public override Activity Tick(Actor self)
+		{
+			self.Trait<IPositionable>().SetPosition(self, destination);
+			self.Generation++;
+			return NextActivity;
+		}
+	}
+}
