+++ OpenRA.Mods.RA/Activities/Teleport.cs
@@ -117,18 +117,4 @@ namespace OpenRA.Mods.RA.Activities
 			return null;
 		}
 	}
-
-	public class SimpleTeleport : Activity
-	{
-		CPos destination;
-
-		public SimpleTeleport(CPos destination) { this.destination = destination; }
-
-		public override Activity Tick(Actor self)
-		{
-			self.Trait<IPositionable>().SetPosition(self, destination);
-			self.Generation++;
-			return NextActivity;
-		}
-	}
 }
